.history/requirements_20250107193223.txt
pandas
pyodbc
requests
beautifulsoup4
boto3
kaggle
pyyaml